+++ src/main/java/com/bookface/postsservice/service/PostsService.java
@@ -7,6 +7,10 @@ import com.bookface.postsservice.firebase.FirebaseInterface;
 import com.bookface.postsservice.model.Post;
 import com.bookface.postsservice.mqconfig.MessagingConfig;
 import com.bookface.postsservice.repository.PostsRepository;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -14,6 +18,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,8 +28,16 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
+import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+import java.net.http.HttpHeaders;
+import java.net.http.HttpRequest.BodyPublishers;
+import java.net.http.HttpResponse.BodyHandlers;
 
 @Service
 @Slf4j
@@ -33,6 +46,8 @@ public class PostsService {
 
     @Autowired
     private final PostsRepository postsRepository;
+
+    private final jwtService jwtService;
 
     @Autowired
     private final CategoriesService categoriesService;
@@ -43,63 +58,110 @@ public class PostsService {
     private final CommentsService commentsService;
     @Autowired
     private RabbitTemplate template;
+    private static final String AUTH_SERVICE_URL = "http://localhost:8080/api/v1/auth/login";
 
+    public boolean compareUsername(HttpServletRequest request, String authorid) {
+        String authorizationHeader = request.getHeader("Authorization");
+        // String encodedCredentials = authorizationHeader.substring(7);
+        String token = authorizationHeader.substring(7);
+        String id = jwtService.extractId(token);
+        return authorid.equals(id);
 
-    public void createPost(PostsRequest postsRequest) throws Exception {
+    }
+
+    public boolean checkIfLoggedInAndToken(HttpServletRequest request) {
+        String authorizationHeader = request.getHeader("Authorization");
+        String token = authorizationHeader.substring(7);
+        boolean isLoggedIn = token != null;
+        if (!isLoggedIn) {
+            return false;
+        } else {
+            boolean checkte = jwtService.isTokenExpired(token);
+            if (checkte) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public String getId(HttpServletRequest request) {
+        String authorizationHeader = request.getHeader("Authorization");
+        String token = authorizationHeader.substring(7);
+        String id = jwtService.extractId(token);
+        return id;
+
+    }
+
+    public void createPost(PostsRequest postsRequest, HttpServletRequest request) throws Exception {
         if (postsRequest.getTitle() == null || postsRequest.getTitle().length() == 0) {
             throw new Exception("Post is missing a Title");
         }
         if (postsRequest.getBody() == null || postsRequest.getBody().length() == 0) {
             throw new Exception("Post cannot be empty");
         }
+// Extract the Base64 encoded credentials from the authorization header
+        String authorizationHeader = request.getHeader("Authorization");
+        String token = authorizationHeader.substring(7);
 
-        //TODO: get authorID from token or mq?
+        boolean isLoggedIn = token != null;
+        if (!isLoggedIn) {
+            throw new Exception("Access denied. User not logged in.");
+        } else {
 
-        Post post = Post.builder()
-                .title(postsRequest.getTitle())
-                .authorId("filler")
-                .body(postsRequest.getBody())
-                .claps(0)
-                .categoryNames(postsRequest.getCategoryNames())
-                .commentCount(0)
-                .published(postsRequest.isPublished())
-                .createdAt(java.time.LocalDateTime.now())
-                .updatedAt(java.time.LocalDateTime.now())
-                .build();
-        //Save image to firebase and save image url.
-        try {
-            List<MultipartFile> files = postsRequest.getFiles();
-            System.out.println(files.size());
-            if(files != null) {
-                for (MultipartFile file : files) {
-                    String fileName = IFirebase.save(file);
-                    String imageUrl = IFirebase.getImageUrl(fileName);
-                    if (post.getPhotoURL() == null) {
-                        post.setPhotoURL(new ArrayList<String>());
+            boolean checkte = jwtService.isTokenExpired(token);
+            if (checkte) {
+                throw new Exception("Token Expired");
+            } else {
+                //TODO: get authorID from token or mq?
+                String username = jwtService.extractId(token);
+                Post post = Post.builder()
+                        .title(postsRequest.getTitle())
+                        .authorId(username)
+                        .body(postsRequest.getBody())
+                        .claps(0)
+                        .categoryNames(postsRequest.getCategoryNames())
+                        .commentCount(0)
+                        .published(postsRequest.isPublished())
+                        .createdAt(java.time.LocalDateTime.now())
+                        .updatedAt(java.time.LocalDateTime.now())
+                        .build();
+                //Save image to firebase and save image url.
+                try {
+                    List<MultipartFile> files = postsRequest.getFiles();
+                    if (files != null) {
+                        for (MultipartFile file : files) {
+                            String fileName = IFirebase.save(file);
+                            String imageUrl = IFirebase.getImageUrl(fileName);
+                            if (post.getPhotoURL() == null) {
+                                post.setPhotoURL(new ArrayList<String>());
+                            }
+
+                            post.getPhotoURL().add(imageUrl);
+                        }
                     }
 
-                    post.getPhotoURL().add(imageUrl);
+                } catch (Exception e) {
+                    System.out.println(e.getMessage());
+                    throw new Exception("Failed to upload Image");
                 }
+
+                postsRepository.insert(post);
+
+                ElasticPostsMessage postMessage = ElasticPostsMessage.builder()
+                        .id(post.getId())
+                        .authorId(post.getAuthorId())
+                        .createdAt(post.getCreatedAt().truncatedTo(ChronoUnit.SECONDS))
+                        .updatedAt(post.getUpdatedAt().truncatedTo(ChronoUnit.SECONDS))
+                        .body(post.getBody())
+                        .title(post.getTitle())
+                        .tags(post.getCategoryNames().toArray(new String[0]))
+                        .build();
+                template.convertAndSend(MessagingConfig.EXCHANGE, MessagingConfig.ROUTING_KEY_CREATE, postMessage);
+
+                log.info("Post {} Saved", post.getId());
             }
 
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-            throw new Exception("Failed to upload Image");
         }
-
-        postsRepository.insert(post);
-
-        ElasticPostsMessage postMessage = ElasticPostsMessage.builder()
-                .id(post.getId())
-                .authorId(post.getAuthorId())
-                .createdAt(post.getCreatedAt().truncatedTo(ChronoUnit.SECONDS))
-                .updatedAt(post.getUpdatedAt().truncatedTo(ChronoUnit.SECONDS))
-                .body(post.getBody())
-                .title(post.getTitle())
-                .build();
-        template.convertAndSend(MessagingConfig.EXCHANGE,MessagingConfig.ROUTING_KEY_CREATE,postMessage);
-
-        log.info("Post {} Saved", post.getId());
     }
 
 
@@ -116,20 +178,48 @@ public class PostsService {
     }
 
 
-    public List<PostsResponse> getPublishedPostsByAuthorId(String authorId) {
+    public List<PostsResponse> getPublishedPostsByAuthorId(HttpServletRequest request) {
+        String authorId = getId(request);
         List<Post> publishedPosts = postsRepository.findByAuthorIdAndPublishedTrue(authorId);
         return publishedPosts.stream().map(this::mapToPostResponse).toList();
     }
-    public List<PostsResponse> getDraftedPostsByAuthorId(String authorId) {
+
+    public List<PostsResponse> getDraftedPostsByAuthorId(HttpServletRequest request) {
+        String authorId = getId(request);
         List<Post> draftedPosts = postsRepository.findByAuthorIdAndPublishedFalse(authorId);
         return draftedPosts.stream().map(this::mapToPostResponse).toList();
     }
 
     @CacheEvict(value = "postCache", key = "#id")
-    public void updatePost(String id, String newTitle, String newBody) throws Exception {
+    public void publishPost(HttpServletRequest request, String id) throws Exception {
+        if (!checkIfLoggedInAndToken(request)) {
+            throw new Exception("User not logged in or token expired");
+        }
+        String authorId = getId(request);
+        Post post = postsRepository.findById(id).orElse(null);
+        if (post != null) {
+            boolean check = compareUsername(request, post.getAuthorId());
+            if (!check) {
+                throw new Exception("Wrong user");
+            }
+            post.setPublished(true);
+            postsRepository.save(post);
+        }
+    }
+
+    @CacheEvict(value = "postCache", key = "#id")
+    public void updatePost(String id, String newTitle, String newBody, HttpServletRequest request) throws Exception {
+        if (!checkIfLoggedInAndToken(request)) {
+            throw new Exception("User not logged in or token expired");
+        }
         Post post = postsRepository.findById(id).orElse(null);
         //TODO: Check if the current session userId matches the posts authorID
         if (post != null) {
+            String username = getId(request);
+            boolean check = compareUsername(request, post.getAuthorId());
+            if (!check) {
+                throw new Exception("Wrong user");
+            }
             if (newTitle != null && newTitle.length() != 0) {
                 post.setTitle(newTitle);
             }
@@ -145,25 +235,33 @@ public class PostsService {
                     .updatedAt(post.getUpdatedAt().truncatedTo(ChronoUnit.SECONDS))
                     .body(post.getBody())
                     .title(post.getTitle())
+                    .tags(post.getCategoryNames().toArray(new String[0]))
                     .build();
-            template.convertAndSend(MessagingConfig.EXCHANGE,MessagingConfig.ROUTING_KEY_UPDATE,postMessage);
-        }
-        else{
+            template.convertAndSend(MessagingConfig.EXCHANGE, MessagingConfig.ROUTING_KEY_UPDATE, postMessage);
+        } else {
             throw new Exception("Post doesn't exist");
         }
     }
 
     @CacheEvict(value = "postCache", key = "#id")
     @Transactional
-    public void deletePost(String id) throws Exception {
+    public void deletePost(String id, HttpServletRequest request) throws Exception {
+        if (!checkIfLoggedInAndToken(request)) {
+            throw new Exception("User not logged in or token expired");
+        }
         Post post = postsRepository.findById(id).orElse(null);
 
         if (post != null) {
-            if(post.getCategoryNames() != null){
+            boolean check = compareUsername(request, post.getAuthorId());
+            if (!check) {
+                throw new Exception("Wrong user");
+
+            }
+            if (post.getCategoryNames() != null) {
                 List<String> categories = post.getCategoryNames();
                 categoriesService.deleteCategoryOrReduceCount(categories);
             }
-            if(post.getPhotoURL() != null) {
+            if (post.getPhotoURL() != null) {
                 for (String imageUrl : post.getPhotoURL()) {
                     try {
                         String path = new URL(imageUrl).getPath();
@@ -177,9 +275,9 @@ public class PostsService {
 
             postsRepository.deleteById(id);
             commentsService.deleteAllCommentsByPostId(id);
-            template.convertAndSend(MessagingConfig.EXCHANGE,MessagingConfig.ROUTING_KEY_DELETE,post.getId());
+            template.convertAndSend(MessagingConfig.EXCHANGE, MessagingConfig.ROUTING_KEY_DELETE, post.getId());
 
-        }else{
+        } else {
             throw new Exception();
         }
     }
@@ -192,7 +290,6 @@ public class PostsService {
             postsRepository.save(post);
         }
     }
-
 
 
     private PostsResponse mapToPostResponse(Post post) {
@@ -211,5 +308,33 @@ public class PostsService {
                 .build();
     }
 
+    private String authenticateUser(String username, String password) throws Exception {
+        // Build the request body with username and password
+        String requestBody = "{\"username\":\"" + username + "\",\"password\":\"" + password + "\"}";
 
+        // Create an HTTP client
+        HttpClient client = HttpClient.newHttpClient();
+
+        // Create an HTTP request with the login API URL
+        HttpRequest request = HttpRequest.newBuilder()
+                .uri(URI.create(AUTH_SERVICE_URL))
+                .header("Content-Type", "application/json")
+                .POST(BodyPublishers.ofString(requestBody))
+                .build();
+
+        HttpResponse<String> response = client.send(request, BodyHandlers.ofString());
+
+        int statusCode = response.statusCode();
+        if (statusCode == 200) {
+            // User is valid, extract the token from the response body
+            String responseBody = response.body();
+            JsonElement jsonElement = JsonParser.parseString(responseBody);
+            JsonObject jsonObject = jsonElement.getAsJsonObject();
+            String token = jsonObject.get("token").getAsString();
+            return token;
+        } else {
+            // User is invalid, handle the error
+            throw new Exception("Invalid username or password");
+        }
+    }
 }
